% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_buff.R
\name{add_buff}
\alias{add_buff}
\title{Add Buff}
\usage{
add_buff(stats, buff)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{buff}{shortcut of the buff to add}
}
\description{
Adds a buff, consumable or worldbuff to stats.
Seed \code{with_buffs} for details.
}
\examples{
stats <- list(int = 275, sp = 581, crit = 5, hit = 9)
add_buff(stats, "ony")
add_buff(stats, "eosp")
}
