% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_consumables.R
\name{add_consumables}
\alias{add_consumables}
\title{Add Consumables}
\usage{
add_consumables(stats, consumables = c("ai", "motw", "gae", "eosp", "bwo"))
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{consumables}{shortcuts of consumables}
}
\description{
Adds consumables and buffs to stats.
Available are:
Arcane Intellect (ai),
Mark of the Wild (motw),
Greater Arcane Elixir (gae),
Elixir of Shadow Power (eosp),
Brilliant Wizard Oil (bwo),
Flask of Supreme Power (fosp)
}
\examples{
add_consumables(list(int = 275, sp = 581, crit = 5, hit = 9))
}
