% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worldbuffs.R
\name{add_worldbuffs}
\alias{add_worldbuffs}
\title{Add Worldbuffs}
\usage{
add_worldbuffs(stats, worldbuffs = c("ony", "zg"))
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{worldbuffs}{shortcuts of worldbuffs}
}
\description{
Adds worldbuffs to stats.
Available are:
Rallying Cry of the Dragonslayer (ony),
Spirit of Zandalar (zg),
Songflower Serenade (ss),
Slip'kik's Savvy (dm)
}
\examples{
add_worldbuffs(list(int = 275, sp = 581, crit = 5, hit = 9))
}
