% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stats.R
\name{clean_stats}
\alias{clean_stats}
\title{Clean Stats}
\usage{
clean_stats(stats)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}
}
\description{
Replaces missing stats with zero.
}
\examples{
clean_stats(list(crit = -1, hit = 1))
}
