% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dps.R
\name{compare_dps}
\alias{compare_dps}
\title{Compare the DPS of different Stats}
\usage{
compare_dps(stats_list, timeframe = c(60, 300), iter = 50000, seed = NULL)
}
\arguments{
\item{stats_list}{as list of \code{stats}}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}

\item{seed}{optional seed}
}
\value{
a data frame
}
\description{
Compare the DPS of different Stats
}
\examples{
compare_dps(list(
  stats1 = list(int = 277, sp = 346, crit = 2, hit = 2),
  stats2 = list(int = 277, sp = 346, mp5 = 100)
),
iter = 1000
)
}
