% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_items.R
\name{compare_items}
\alias{compare_items}
\title{Compare Two Items}
\usage{
compare_items(
  crit,
  hit,
  int,
  sp,
  mp5,
  d_crit = 0,
  d_hit = 0,
  d_int = 0,
  d_sp = 0,
  d_mp5 = 0,
  timeframe = c(45, 150),
  iter = 50000
)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{mp5}{mp5 on gear}

\item{d_crit}{difference in crit}

\item{d_hit}{difference in hit}

\item{d_int}{difference in intelligence}

\item{d_sp}{difference in spellpower}

\item{d_mp5}{difference in mp5}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}
}
\value{
a vector with current dps, new dps and difference
}
\description{
Simulates dps with current stats and with new stats to compare them.
}
\examples{
compare_items(2, 1, 277, 346, 0, -1, 1, 5, 3, 0, iter = 1000)
}
