% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_critchance.R
\name{compute_critchance}
\alias{compute_critchance}
\title{Compute Crit Chance}
\usage{
compute_critchance(crit, int, devastation = 5)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{int}{intellect of character}

\item{devastation}{number of talent points in devastation}
}
\description{
Computes crit chance based on crit on gear, intelligence and talent points in
devastation.
}
\examples{
compute_critchance(3, 267)
}
