% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mana.R
\name{compute_mana}
\alias{compute_mana}
\title{Compute Mana}
\usage{
compute_mana(int, base = 1093)
}
\arguments{
\item{int}{intellect of character}

\item{base}{base mana}
}
\description{
Computes the maximum mana based on intellect and base mana.
}
\examples{
compute_mana(267)
}
