% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_manacost.R
\name{compute_manacost}
\alias{compute_manacost}
\title{Compute Mana Cost}
\usage{
compute_manacost(cataclysm = 2)
}
\arguments{
\item{cataclysm}{number of talent points in cataclysm}
}
\description{
Computes the mana cost of a shadow bolt based on numer of points in cataclysm.
}
\examples{
compute_manacost()
}
