% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_miss.R
\name{compute_miss}
\alias{compute_miss}
\title{Compute Miss}
\usage{
compute_miss(s, hit)
}
\arguments{
\item{s}{random numbers returned by \code{sample_roll}}

\item{hit}{hit percentage on gear}
}
\value{
logical vector
}
\description{
Determines from the sample whether a spell misses.
}
\examples{
compute_miss(sample_roll(10), 2)
}
