% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_stats.R
\name{diff_stats}
\alias{diff_stats}
\title{Diff Stats}
\usage{
diff_stats(x, y)
}
\arguments{
\item{x, y}{stats to take the difference \code{x - y}}
}
\description{
Diff Stats
}
\examples{
diff_stats(list(int = 5, sp = 2, crit = 1), list(int = 2, sp = 3, hit = 1))
}
