% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_call_stats.R
\name{do_call_stats}
\alias{do_call_stats}
\title{Execute a Stat Function Call}
\usage{
do_call_stats(args, what)
}
\arguments{
\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{what}{either a function or a non-empty character string naming the
    function to be called.}
}
\description{
This is a wrapper around \code{do.call} for stat input taking care of cleaning and
formatting.
}
