% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_equip.R
\name{is_equip}
\alias{is_equip}
\title{Is Equip}
\usage{
is_equip(equip)
}
\arguments{
\item{equip}{equipment list}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if an equipment list has all slots filled.
These are head, neck, shoulders, back, chest, wrist, hands, waist, legs,
feet, finger1, finger2, trinket1, trinket2 and weapon.
Every slot should be filled with a list of the item base stats.
}
