% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stats.R
\name{is_stats}
\alias{is_stats}
\title{Is Stats}
\usage{
is_stats(stats)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}
}
\description{
Checks if an object can be interpreted as stats.
}
