% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetap_impl.R
\name{lifetap_impl}
\alias{lifetap_impl}
\title{Base Implementation of Life Tap}
\usage{
lifetap_impl(manacost, improved_sb_proc)
}
\arguments{
\item{manacost}{mana cost}

\item{improved_sb_proc}{number of active procs of improved shadow bolt on
target}
}
\description{
Base Implementation of Life Tap
}
