% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetap_mana.R
\name{lifetap_mana}
\alias{lifetap_mana}
\title{Life Tap Mana}
\usage{
lifetap_mana(sp, talents = warlock_talents())
}
\arguments{
\item{sp}{(shadow) spellpower of character}

\item{talents}{list of talents}
}
\description{
Life Tap Mana
}
\examples{
lifetap_mana(0)
}
