% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetap_manacost.R
\name{lifetap_manacost}
\alias{lifetap_manacost}
\title{Life Tap Mana Cost}
\usage{
lifetap_manacost(sp, imp_lifetap = 2)
}
\arguments{
\item{sp}{(shadow) spellpower on gear}

\item{imp_lifetap}{number of talent points in improved life tap}
}
\description{
Life Tap Mana Cost
}
\examples{
lifetap_manacost(0)
}
