% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_buffs.R
\name{rank_buffs}
\alias{rank_buffs}
\title{Rank Buffs}
\usage{
rank_buffs(
  stats,
  consumables = character(),
  worldbuffs = character(),
  timeframe = c(60, 300),
  iter = 50000,
  seed = NULL
)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{consumables}{shortcuts of consumables}

\item{worldbuffs}{shortcuts of worldbuffs}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}

\item{seed}{optional seed}
}
\description{
Ranks buffs by choosing the best one iteratively.
}
\examples{
stats <- with_buffs(list(int = 275, sp = 581, crit = 5, hit = 9))
rank_buffs(stats, consumables = c("gae", "eosp", "bwo"), iter = 1000)
rank_buffs(stats, worldbuffs = c("ony", "zg", "ss", "dm"), iter = 1000)
}
