% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_all.R
\name{sample_all}
\alias{sample_all}
\title{Sample Shadow Bolt Damage, Hit, Crit and Curse Hit}
\usage{
sample_all(n)
}
\arguments{
\item{n}{number of samples}
}
\value{
a list
}
\description{
Sample Shadow Bolt Damage, Hit, Crit and Curse Hit
}
\examples{
sample_all(10)
}
