% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_hit.R
\name{sample_hit}
\alias{sample_hit}
\title{Sample Hit}
\usage{
sample_hit(n = 1)
}
\arguments{
\item{n}{number of samples}
}
\description{
Vectorised function to simulate wheather a spell hits or crits.
}
\examples{
sample_hit()
sample_hit(5)
}
