% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_roll.R
\name{sample_roll}
\alias{sample_roll}
\title{Sample Roll}
\usage{
sample_roll(n = 1)
}
\arguments{
\item{n}{number of samples}
}
\description{
Vectorised function to simulate whether a spell hits or crits.
}
\examples{
sample_roll()
sample_roll(5)
}
