% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_shadowbolt.R
\name{sample_shadowbolt}
\alias{sample_shadowbolt}
\title{Sample Shadow Bolt}
\usage{
sample_shadowbolt(n = 1)
}
\arguments{
\item{n}{number of samples}
}
\description{
Use the sample function so simulate the base damage of a shadow bolt.
Intended for vectorisation if the simulation.
}
\examples{
sample_shadowbolt()
sample_shadowbolt(5)
}
