% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sim.R
\name{sample_sim}
\alias{sample_sim}
\title{Sample Shadow Bolt Damage, Hit, Crit and Curse Hit}
\usage{
sample_sim(n)
}
\arguments{
\item{n}{number of samples}
}
\value{
a list
}
\description{
Sample Shadow Bolt Damage, Hit, Crit and Curse Hit
}
