% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt.R
\name{shadowbolt}
\alias{shadowbolt}
\title{Shadow Bolt}
\usage{
shadowbolt(
  crit,
  hit,
  int,
  sp,
  talents = warlock_talents(),
  improved_sb_proc = 0
)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower of character}

\item{talents}{list of talents}

\item{improved_sb_proc}{number of active procs of improved shadow bolt on
target}
}
\value{
a numeric vector with values for dmg, mana, casttime and
improved_sb_proc
}
\description{
Simulates a shadow bolt.
}
\examples{
shadowbolt(2, 1, 277, 346)
}
