% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt_dmg.R
\name{shadowbolt_dmg}
\alias{shadowbolt_dmg}
\title{Compute Shadow Bolt Damage}
\usage{
shadowbolt_dmg(
  sample_dmg,
  sp,
  sb_miss,
  sb_crit,
  devastation = 5,
  ruin = 1,
  shadow_mastery = 0,
  demonic_sacrifice = 1,
  curse_of_shadows = 1
)
}
\arguments{
\item{sample_dmg}{base damage returned by \code{sample_shadowbolt}}

\item{sp}{(shadow) spellpower on gear}

\item{sb_miss}{vector of whether shadow bolt hits}

\item{sb_crit}{vector of whether shadow bolt crits}

\item{devastation}{number of talent points in devastation}

\item{ruin}{number of talent points in ruin}

\item{shadow_mastery}{number of talent points in shadow mastery}

\item{demonic_sacrifice}{whether succubus was sacrificed}

\item{curse_of_shadows}{whether curse of shadows is on target}
}
\description{
Compute Shadow Bolt Damage
}
