% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt_impl.R
\name{shadowbolt_impl}
\alias{shadowbolt_impl}
\title{Base Implementation of Shadow Bolt}
\usage{
shadowbolt_impl(sb_dmg, sb_miss, sb_crit, manacost, casttime, improved_sb_proc)
}
\arguments{
\item{sb_dmg}{vector of base shadow bolt damage}

\item{sb_miss}{vector of whether shadow bolt hits}

\item{sb_crit}{vector of whether shadow bolt crits}

\item{manacost}{mana cost}

\item{casttime}{cast time}

\item{improved_sb_proc}{number of active procs of improved shadow bolt on
target}
}
\description{
Base Implementation of Shadow Bolt
}
\examples{
shadowbolt_impl(575, FALSE, TRUE, 363, 2.5, 0)
}
