% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt_mana.R
\name{shadowbolt_mana}
\alias{shadowbolt_mana}
\title{Shadow Bolt Mana}
\usage{
shadowbolt_mana(talents = warlock_talents())
}
\arguments{
\item{talents}{list of talents}
}
\description{
Shadow Bolt Mana
}
