% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt_manacost.R
\name{shadowbolt_manacost}
\alias{shadowbolt_manacost}
\title{Shadow Bolt Mana Cost}
\usage{
shadowbolt_manacost(cataclysm = 2)
}
\arguments{
\item{cataclysm}{number of talent points in cataclysm}
}
\description{
Shadow Bolt Mana Cost
}
\examples{
shadowbolt_manacost()
}
