% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_boss_impl.R
\name{sim_boss_impl}
\alias{sim_boss_impl}
\title{Base Implementation Simulate Boss Fight}
\usage{
sim_boss_impl(
  mana,
  mp5,
  sb_dmg,
  sb_miss,
  sb_crit,
  sb_manacost,
  lt_mancost,
  time
)
}
\arguments{
\item{mana}{starting mana (by default computed from int)}

\item{mp5}{mp5 on gear}

\item{sb_dmg}{vector of shadow bolt damage ignoring possible miss or crit}

\item{sb_miss}{vector of whether shadow bolt hits}

\item{sb_crit}{vector of whether shadow bolt crits}

\item{sb_manacost}{mana cost of shadow bolt}

\item{lt_mancost}{mana cost of life tap}

\item{time}{length of fight in seconds}
}
\description{
Base Implementation Simulate Boss Fight
}
