% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_boss_loop}
\alias{sim_boss_loop}
\title{Simulate the Loop for a Boss Fight}
\usage{
sim_boss_loop(
  mana,
  mp5,
  sb_dmg,
  sb_miss,
  sb_crit,
  curse_miss,
  sb_manacost,
  lt_manacost,
  time
)
}
\arguments{
\item{mana}{starting mana}

\item{mp5}{mp5 on gear}

\item{sb_dmg}{vector of base shadow bolt damage}

\item{sb_miss}{vector of whether shadow bolt hits}

\item{sb_crit}{vector of whether shadow bolt crits}

\item{curse_miss}{vector of whether curse hits}

\item{sb_manacost}{mana cost of shadow bolt}

\item{lt_manacost}{mana cost of life tap}

\item{time}{length of fight in seconds}
}
\description{
Simulate the Loop for a Boss Fight
}
