% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dps.R
\name{sim_dps}
\alias{sim_dps}
\title{Simultate DPS}
\usage{
sim_dps(
  stats,
  timeframe = c(60, 300),
  iter = 50000,
  seed = NULL,
  trinkets = NULL
)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}

\item{seed}{optional seed}

\item{trinkets}{which use trinkets are equipped, currently implemented:
toep (Talisman of Ephemeral Power)}
}
\value{
a matrix where every row is a vector returned by \code{sim_boss}
}
\description{
Repeatedly simulates bosses to calculate dps.
The timeframe sets a minimum an maximum time for a boss fight.
Each boss simulation has a random length in between.
}
\examples{
mat <- sim_dps(list(int = 277, sp = 346, crit = 2, hit = 2), iter = 1000)
mean(mat[, 4])
}
