% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{sim_setup}
\alias{sim_setup}
\title{Setup Simulation}
\usage{
sim_setup(n, crit, hit, int, sp, mana, iter = 1)
}
\arguments{
\item{n}{sample number}

\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{mana}{starting mana (by default computed from int)}

\item{iter}{number of iterations}
}
\value{
a list of arguments
}
\description{
Setup Simulation
}
\examples{
sim_setup(10, 2, 1, 277, 346, NULL)
}
