% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{Sum Stats}
\usage{
sum_stats(stats_list)
}
\arguments{
\item{stats_list}{as list of \code{stats}}
}
\value{
\code{stats}
}
\description{
Sum Stats
}
\examples{
sum_stats(list(
  list(int = 1, sp = 2),
  list(crit = 3, hit = 4)
))
}
