% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinket_sp.R
\name{trinket_sp}
\alias{trinket_sp}
\title{Spell Power Bonus of Trinkets}
\usage{
trinket_sp(trinkets, times)
}
\arguments{
\item{trinkets}{which use trinkets are equipped, currently implemented:
toep (Talisman of Ephemeral Power)
zhc (Zandalarian Hero Charm)}

\item{times}{vector with \code{time} for every iteration}
}
\description{
Spell Power Bonus of Trinkets
}
\examples{
trinket_sp("toep", 150:154)
trinket_sp(c("toep", "zhc"), 150)
}
