% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warlock_talents.R
\name{warlock_talents}
\alias{warlock_talents}
\title{Warlock Talents}
\usage{
warlock_talents(
  devastation = 5,
  ruin = 1,
  improved_sb = 5,
  cataclysm = 2,
  bane = 5,
  shadow_mastery = 0,
  demonic_sacrifice = 1,
  imp_lifetap = 2,
  suppression = 2,
  curse_of_shadows = 1
)
}
\arguments{
\item{devastation}{number of talent points in devastation}

\item{ruin}{number of talent points in ruin}

\item{improved_sb}{number of talent points in improved shadow bolt}

\item{cataclysm}{number of talent points in cataclysm}

\item{bane}{number of talent points in bane}

\item{shadow_mastery}{number of talent points in shadow mastery}

\item{demonic_sacrifice}{whether succubus was sacrificed}

\item{imp_lifetap}{number of talent points in improved life tap}

\item{suppression}{number of talent points in suppression}

\item{curse_of_shadows}{whether curse of shadows is on target}
}
\description{
Creates a list with all the relevant warlock talents.
The default is the standard built.
}
\examples{
warlock_talents()
}
