% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warlockr-package.R
\docType{package}
\name{warlockr-package}
\alias{warlockr}
\alias{warlockr-package}
\title{warlockr: Classic WoW Warlock Related Stuff}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cphaarmeyer/warlockr}
  \item Report bugs at \url{https://github.com/cphaarmeyer/warlockr/issues}
}

}
\keyword{internal}
