% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_buffs.R
\name{with_buffs}
\alias{with_buffs}
\title{With Buffs}
\usage{
with_buffs(
  stats,
  buffs = c("ai", "motw"),
  consumables = character(),
  worldbuffs = character()
)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{buffs}{shortcuts of buffs}

\item{consumables}{shortcuts of consumables}

\item{worldbuffs}{shortcuts of worldbuffs}
}
\description{
Adds buffs to stats.
}
\details{
Available buffs:
Arcane Intellect (ai),
Mark of the Wild (motw)

Available consumables:
Greater Arcane Elixir (gae),
Elixir of Shadow Power (eosp),
Brilliant Wizard Oil (bwo),
Flask of Supreme Power (fosp)

Available worldbuffs:
Rallying Cry of the Dragonslayer (ony),
Spirit of Zandalar (zg),
Songflower Serenade (ss),
Slip'kik's Savvy (dm)
}
\examples{
with_buffs(list(int = 275, sp = 581, crit = 5, hit = 9),
  worldbuffs = c("ony", "zg")
)
}
