% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_buffs.R
\name{compare_buffs}
\alias{compare_buffs}
\title{Compare Buffs}
\usage{
compare_buffs(stats, timeframe = c(60, 300), iter = 50000)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}
}
\value{
a data frame
}
\description{
Simulates dps with current stats and with new stats to compare them.
}
\examples{
compare_buffs(with_buffs(
  list(int = 275, sp = 581, crit = 5, hit = 9),
  buffs = c("ai", "motw")
),
iter = 1000
)
}
