% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_items.R
\name{compare_items}
\alias{compare_items}
\title{Compare Items}
\usage{
compare_items(stats, items, timeframe = c(60, 300), iter = 50000)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{items}{list where each element indicates the stat changes by an item}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}
}
\value{
a data frame
}
\description{
Simulates dps with current stats and with new stats to compare them.
}
\examples{
compare_items(
  list(int = 277, sp = 346, crit = 2, hit = 2),
  items = list(
    new_item1 = list(crit = -1, hit = 1),
    new_item2 = list(int = 5, sp = 3)
  ),
  iter = 1000
)
}
