% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_items.R
\name{compare_items}
\alias{compare_items}
\title{Compare Two Items}
\usage{
compare_items(
  crit,
  hit,
  int,
  sp,
  mp5,
  items,
  timeframe = c(45, 150),
  iter = 50000
)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{mp5}{mp5 on gear}

\item{items}{list where each element indicates the stat changes by an item}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}
}
\value{
a data frame with current dps, new dps and difference
}
\description{
Simulates dps with current stats and with new stats to compare them.
}
\examples{
compare_items(2, 1, 277, 346, 0,
  items = list(
    new_item1 = list(crit = -1, hit = 1),
    new_item2 = list(int = 5, sp = 3)
  ),
  iter = 1000
)
}
