% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_manacost.R
\name{compute_manacost}
\alias{compute_manacost}
\title{Compute Mana Cost}
\usage{
compute_manacost(
  spell = c("shadowbolt", "lifetap"),
  cataclysm = 2,
  sp = 0,
  imp_lifetap = 2
)
}
\arguments{
\item{spell}{which spell}

\item{cataclysm}{number of talent points in cataclysm}

\item{sp}{(shadow) spellpower on gear}

\item{imp_lifetap}{number of talent points in improved life tap}
}
\description{
Computes the mana cost of a spell.
}
\examples{
compute_manacost()
}
