% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_statweights.R
\name{compute_statweights}
\alias{compute_statweights}
\title{Compute Stat Weights}
\usage{
compute_statweights(
  stats,
  timeframe = c(60, 300),
  iter = 50000,
  seed = NULL,
  trinkets = NULL
)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}

\item{seed}{optional seed}

\item{trinkets}{which use trinkets are equipped, currently implemented:
toep (Talisman of Ephemeral Power)
zhc (Zandalarian Hero Charm)}
}
\value{
a data frame with weights
}
\description{
Simulates current dps and dps with increased stats to compute weights for
these stats.
The result is scaled such that spell power has weight 1.
}
\examples{
compute_statweights(
  list(int = 277, sp = 346, crit = 2, hit = 2),
  iter = 1000
)
}
