% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_statweights.R
\name{compute_statweights}
\alias{compute_statweights}
\title{Compute Stat Weights}
\usage{
compute_statweights(
  crit,
  hit,
  int,
  sp,
  mp5,
  timeframe = c(45, 150),
  iter = 50000
)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{mp5}{mp5 on gear}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}
}
\value{
a named vector with weights
}
\description{
Simulates current dps and dps with inceased stats to compute weights for these stats.
}
\examples{
compute_statweights(2, 1, 277, 346, 0, iter = 1000)
}
