% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_statweights.R
\name{compute_statweights}
\alias{compute_statweights}
\title{Compute Stat Weights}
\usage{
compute_statweights(stats, timeframe = c(45, 150), iter = 50000)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp, crit, hit and mp5}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}
}
\value{
a named vector with weights
}
\description{
Simulates current dps and dps with inceased stats to compute weights for these stats.
}
\examples{
compute_statweights(list(int = 277, sp = 346, crit = 2, hit = 2), iter = 1000)
}
