% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetap.R
\name{lifetap}
\alias{lifetap}
\title{Life Tap}
\usage{
lifetap(sp, imp_lifetap = 2, improved_sb_proc = 0)
}
\arguments{
\item{sp}{(shadow) spellpower of character}

\item{imp_lifetap}{number of talent points in improved life tap}

\item{improved_sb_proc}{number of active procs of improved shadow bolt on
target}
}
\value{
a numeric vector with values for dmg, manacost, casttime and
improved_sb_proc
}
\description{
Simulates a life tap.
}
\examples{
lifetap(346)
}
