% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt.R
\name{shadowbolt}
\alias{shadowbolt}
\title{Shadowbolt}
\usage{
shadowbolt(
  crit,
  hit,
  int,
  sp,
  devastation = 5,
  ruin = 1,
  improved_sb = 5,
  cataclysm = 2,
  bane = 5,
  shadow_mastery = 0,
  demonic_sacrifice = 1,
  improved_sb_proc = 0,
  curse_of_shadows = 1
)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{devastation}{number of talent points in devastation}

\item{ruin}{number of talent points in ruin}

\item{improved_sb}{number of talent points in improved shadow bolt}

\item{cataclysm}{number of talent points in cataclysm}

\item{bane}{number of talent points in bane}

\item{shadow_mastery}{number of talent points in shadow mastery}

\item{demonic_sacrifice}{whether succubus was sacrificed}

\item{improved_sb_proc}{number of active procs of improved shadow bolt on target}

\item{curse_of_shadows}{whether curse of shadows is on target}
}
\value{
a numeric vector with values for dmg, manacost, casttime and improved_sb_proc
}
\description{
Simulates a shadowbolt. Based on:
https://www.reddit.com/r/classicwow/comments/dh5r6g/so_i_made_a_warlock_shadowbolt_simulator/
}
\examples{
shadowbolt(2, 1, 277, 346)
}
