% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt_dmg.R
\name{shadowbolt_dmg}
\alias{shadowbolt_dmg}
\title{Compute Shadow Bolt Damage}
\usage{
shadowbolt_dmg(
  sb_dmg,
  sp,
  sb_miss,
  sb_crit,
  ruin = 1,
  improved_sb = 5,
  shadow_mastery = 0,
  demonic_sacrifice = 1,
  curse_of_shadows = 1,
  improved_sb_proc = 0
)
}
\arguments{
\item{sb_dmg}{vector of base shadow bolt damage}

\item{sp}{(shadow) spellpower on gear}

\item{sb_miss}{vector of whether shadow bolt hits}

\item{sb_crit}{vector of whether shadow bolt crits}

\item{ruin}{number of talent points in ruin}

\item{improved_sb}{number of talent points in improved shadow bolt}

\item{shadow_mastery}{number of talent points in shadow mastery}

\item{demonic_sacrifice}{whether succubus was sacrificed}

\item{curse_of_shadows}{whether curse of shadows is on target}

\item{improved_sb_proc}{number of active procs of improved shadow bolt on
target}
}
\description{
Compute Shadow Bolt Damage
}
