% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt_impl.R
\name{shadowbolt_impl}
\alias{shadowbolt_impl}
\title{Base Implementation of Shadow Bolt}
\usage{
shadowbolt_impl(
  dmg,
  miss_test,
  crit_test,
  manacost,
  casttime,
  improved_sb,
  improved_sb_proc
)
}
\arguments{
\item{dmg}{damage ignoring improved shadow bolt proc}

\item{miss_test}{whether hits}

\item{crit_test}{whether crits}

\item{manacost}{mana cost}

\item{casttime}{cast time}

\item{improved_sb}{number of talent points in improved shadow bolt}

\item{improved_sb_proc}{number of active procs of improved shadow bolt on target}
}
\description{
Base Implementation of Shadow Bolt
}
\examples{
shadowbolt_impl(575, FALSE, TRUE, 363, 2.5, 5, 0)
}
