% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowbolt_mana.R
\name{shadowbolt_mana}
\alias{shadowbolt_mana}
\title{Shadow Bolt Mana}
\usage{
shadowbolt_mana(cataclysm = 2)
}
\arguments{
\item{cataclysm}{number of talent points in cataclysm}
}
\description{
Shadow Bolt Mana
}
\examples{
shadowbolt_mana()
}
