% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_boss.R
\name{sim_boss}
\alias{sim_boss}
\title{Simulate Boss Fight}
\usage{
sim_boss(crit, hit, int, sp, mana, time)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{mana}{starting mana}

\item{time}{length of fight in seconds}
}
\value{
a numeric vector with values for total dmg, mana, total casttime and dps
}
\description{
Simulates one boss fight with default values. Based on:
https://www.reddit.com/r/classicwow/comments/dh5r6g/so_i_made_a_warlock_shadowbolt_simulator/
}
\examples{
sim_boss(2, 1, 277, 346, 5300, 150)
}
