% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_boss.R
\name{sim_boss}
\alias{sim_boss}
\title{Simulate Boss Fight}
\usage{
sim_boss(stats, time = 150)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp, crit, hit and mp5}

\item{time}{length of fight in seconds}
}
\value{
a numeric vector with values for total dmg, mana, total casttime and dps
}
\description{
Simulates one boss fight with default values.
}
\examples{
sim_boss(list(int = 277, sp = 346, crit = 2, hit = 2))
}
