% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_boss_impl.R
\name{sim_boss_impl}
\alias{sim_boss_impl}
\title{Base Implementation Simulate Boss Fight}
\usage{
sim_boss_impl(
  mana,
  mp5,
  sp,
  sb_dmg,
  sb_miss,
  sb_crit,
  curse_miss,
  sb_mana,
  lt_mana,
  time,
  sp_bonus = NULL
)
}
\arguments{
\item{mana}{starting mana}

\item{mp5}{mp5 on gear}

\item{sp}{(shadow) spellpower of character}

\item{sb_dmg}{vector of base shadow bolt damage}

\item{sb_miss}{vector of whether shadow bolt hits}

\item{sb_crit}{vector of whether shadow bolt crits}

\item{curse_miss}{vector of whether curse hits}

\item{sb_mana}{mana difference of shadow bolt}

\item{lt_mana}{mana difference of life tap}

\item{time}{length of fight in seconds}

\item{sp_bonus}{list of sp bonus vectors applied to shadow bolts}
}
\description{
Base Implementation Simulate Boss Fight
}
