% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dps.R
\name{sim_dps}
\alias{sim_dps}
\title{Simultate DPS}
\usage{
sim_dps(
  crit,
  hit,
  int,
  sp,
  mp5,
  timeframe = c(45, 150),
  iter = 50000,
  seed = NULL
)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{mp5}{mp5 on gear}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}

\item{seed}{optional seed}
}
\value{
a matrix where every row is a vector returned by \code{sim_boss}
}
\description{
Repeatedly simulates bosses to calculate dps.
The timeframe sets a minimum an maximum time for a boss fight.
Each boss simulation has a random length in between.
}
\examples{
mat <- sim_dps(2, 1, 277, 346, 0, iter = 1000)
mean(mat[, 4])
}
