% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dps.R
\name{sim_dps}
\alias{sim_dps}
\title{Simultate DPS}
\usage{
sim_dps(crit, hit, int, sp, mana = NULL, timeframe = c(45, 150), iter = 50000)
}
\arguments{
\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{mana}{starting mana (by default computed from int)}

\item{timeframe}{vector of minimal and maximal duration of a fight in seconds}

\item{iter}{number of iterations}
}
\value{
a matrix where every row is a vector returned by \code{sim_boss}
}
\description{
Repeatedly simulates bosses to calculate dps.
}
\examples{
mat <- sim_dps(2, 1, 277, 346, iter = 1000)
mean(mat[, 4])
}
