% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{sim_setup}
\alias{sim_setup}
\title{Setup Simulation}
\usage{
sim_setup(
  times,
  crit,
  hit,
  int,
  sp,
  iter = 1,
  devastation = 5,
  cataclysm = 2,
  suppression = 2
)
}
\arguments{
\item{times}{vector with \code{time} for every iteration}

\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{iter}{number of iterations}

\item{devastation}{number of talent points in devastation}

\item{cataclysm}{number of talent points in cataclysm}

\item{suppression}{number of talent points in suppression}
}
\value{
a list of arguments
}
\description{
Setup Simulation
}
\examples{
sim_setup(30, 2, 1, 277, 346)
}
