% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{sim_setup}
\alias{sim_setup}
\title{Setup Simulation}
\usage{
sim_setup(
  times,
  crit,
  hit,
  int,
  sp,
  trinkets = NULL,
  iter = 1,
  talents = warlock_talents()
)
}
\arguments{
\item{times}{vector with \code{time} for every iteration}

\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower of character}

\item{trinkets}{which use trinkets are equipped, currently implemented:
toep (Talisman of Ephemeral Power)
zhc (Zandalarian Hero Charm)}

\item{iter}{number of iterations}

\item{talents}{list of talents}
}
\value{
a list of arguments
}
\description{
Setup Simulation
}
