% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{sim_setup}
\alias{sim_setup}
\title{Setup Simulation}
\usage{
sim_setup(
  n,
  crit,
  hit,
  int,
  sp,
  iter = 1,
  devastation = 5,
  ruin = 1,
  cataclysm = 2,
  shadow_mastery = 0,
  demonic_sacrifice = 1,
  curse_of_shadows = 1,
  suppression = 2
)
}
\arguments{
\item{n}{sample number}

\item{crit}{crit percentage on gear}

\item{hit}{hit percentage on gear}

\item{int}{intellect of character}

\item{sp}{(shadow) spellpower on gear}

\item{iter}{number of iterations}

\item{devastation}{number of talent points in devastation}

\item{ruin}{number of talent points in ruin}

\item{cataclysm}{number of talent points in cataclysm}

\item{shadow_mastery}{number of talent points in shadow mastery}

\item{demonic_sacrifice}{whether succubus was sacrificed}

\item{curse_of_shadows}{whether curse of shadows is on target}

\item{suppression}{number of talent points in suppression}
}
\value{
a list of arguments
}
\description{
Setup Simulation
}
\examples{
sim_setup(10, 2, 1, 277, 346)
}
