% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_buffs.R
\name{with_buffs}
\alias{with_buffs}
\title{With Buffs}
\usage{
with_buffs(
  stats,
  buffs = c("ai", "motw"),
  consumables = c("gae", "eosp", "bwo"),
  worldbuffs = c("ony", "zg")
)
}
\arguments{
\item{stats}{a named list with the stats of the character including int, sp,
crit, hit and mp5}

\item{buffs}{shortcuts of buffs}

\item{consumables}{shortcuts of consumables}

\item{worldbuffs}{shortcuts of worldbuffs}
}
\description{
Adds buffs to stats.
}
\examples{
with_buffs(list(int = 275, sp = 581, crit = 5, hit = 9))
}
